// express-backend/src/db/schema.ts
export type Generated<T> = T extends ColumnType<infer S, infer I, infer U>
  ? ColumnType<S, I | undefined, U>
  : ColumnType<T, T | undefined, T>;

export type Int8 = ColumnType<number, bigint | number | string, bigint | number | string>;

export type Json = JsonValue;

export type JsonArray = JsonValue[];

export type JsonObject = {
  [x: string]: JsonValue | undefined;
};

export type JsonPrimitive = boolean | number | string | null;

export type JsonValue = JsonArray | JsonObject | JsonPrimitive;

export type Numeric = ColumnType<string, number | string, number | string>;

export type Timestamp = ColumnType<string, Date | string, Date | string>;
// ...
export interface LeaseagreementsLeaseacceptance {
  accepted_at: Timestamp;
  booking_id: Int8;
  id: Generated<Int8>;
  ip_address: string | null;
  user_id: Int8;
  version: string;
}

export interface LeaseagreementsLeaseagreement {
  booking_id: Int8;
  document: string | null;
  generated_at: Timestamp;
  id: Generated<Int8>;
  landlord_name: string;
  lease_end: Timestamp | null;
  lease_start: Timestamp | null;
  property_address: string;
  rent_display: string;
  tenant_name: string;
  updated_at: Timestamp;
  version: string;
}